css: contrast ratios
The background for dark mode is pitch black for OLED screens now. Not
sure if this is the best idea, or not though...

The light mode has lighter backgrounds and better contrast between text
and the background. Tools still complain about shit contrast, but I
can't seem to pick colours that are more contrasty and also seem
different enough from the regular text…